Move logic away from routes
